pyproject.toml
[tool.poetry]
name = "datapizza-ai-test"
version = "0.1.0"
description = ""
authors = ["alessiogandelli <alessiogandelli99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
pandas = "^2.3.3"
dotenv = "^0.9.9"
datapizza-ai = "^0.0.2"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
 